test: fix invalid date provider to be static in PayloadTransformerInvalidValuesTest
