+++ backend/tests/Services/PayloadTransformerRequiredFieldsTest.php
@@ -7,14 +7,21 @@ use App\Services\PayloadTransformer;
 /**
  * @covers \App\Services\PayloadTransformer
  */
-class PayloadTransformerMissingFieldsTest extends TestCase
+class PayloadTransformerRequiredFieldsTest extends TestCase
 {
+    private PayloadTransformer $transformer;
+
+    protected function setUp(): void
+    {
+        $this->transformer = new PayloadTransformer();
+    }
+
     public function testThrowsExceptionForMissingArrival(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         $this->expectExceptionMessage("Missing required field: Arrival");
 
-        (new PayloadTransformer())->transform([
+        $this->transformer->transform([
             "Departure" => "05/10/2025",
             "Ages"      => [25, 12]
         ]);
@@ -25,7 +32,7 @@ class PayloadTransformerMissingFieldsTest extends TestCase
         $this->expectException(\InvalidArgumentException::class);
         $this->expectExceptionMessage("Missing required field: Departure");
 
-        (new PayloadTransformer())->transform([
+        $this->transformer->transform([
             "Arrival" => "01/10/2025",
             "Ages"    => [25, 12]
         ]);
@@ -36,7 +43,7 @@ class PayloadTransformerMissingFieldsTest extends TestCase
         $this->expectException(\InvalidArgumentException::class);
         $this->expectExceptionMessage("Missing required field: Ages");
 
-        (new PayloadTransformer())->transform([
+        $this->transformer->transform([
             "Arrival"   => "01/10/2025",
             "Departure" => "05/10/2025"
         ]);
